Redirections and other fixes

+++ parser/parsing/parse_config_file.cpp
@@ -108,7 +108,7 @@ std::vector<ServerData>	read_config_file(std::string file)
 			ServerData server;
 
 			server.server_name = server_ptr -> server_name.get<std::string>("default");
-			server.port_to_listen = server_ptr -> listen.get<short>(80);
+			server.port_to_listen = server_ptr -> listen.get<int>(80);
 			server.root = server_ptr -> root.get<std::string>("/");
 			server.index_file = server_ptr -> index.get<std::string>("test");
